# static/swagger.yml
openapi: 3.0.0
paths:
...
  /api/v2/file:
    post:
      operationId: files-get
      summary: Получение списка файлов
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FilesGetRequest"
      responses:
        "200":
          description: Успешный ответ
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FilesGetResponse"
        "400": &ref_32
          description: Ответ будет таким если с данным что-то не так
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BadRequestError"
        "401": &ref_33
          description: Ответ для незарегистрированного пользователя
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnauthorizedError"
        "403": &ref_34
          description: Ответ для неавторизованного пользователя
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ForbiddenError"
        "404": &ref_35
          description: Не найдено
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotFoundError"
        "406": &ref_36
          description: Не принято значение
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotAcceptableError"
        "409": &ref_37
          description: Ответ для конфликта файлов
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConflictError"
        "412": &ref_38
          description: Пользователь уже существует
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PreconditionFailedError"
        "500": &ref_39
          description: Ошибка сервера
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InternalServerError"
        "501": &ref_40
          description: Пока не реализовано
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotImplementedError"
      tags: &ref_41
        - file
      security: &ref_42
        - bearer: []
    put:
      operationId: file-upload
      summary: Загрузка файлов
      parameters: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - files
                - param
              properties:
                files:
                  type: array
                  description: Файл(ы)
                  items:
                    type: string
                    format: binary
                param:
                  type: object
                  description: Параметры загрузки файла
                  $ref: "#/components/schemas/FileUploadRequest"
      responses:
        "200":
          description: Успешный ответ
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FilesUploadResponse"
        "400": *ref_32
        "401": *ref_33
        "403": *ref_34
        "404": *ref_35
        "406": *ref_36
        "409": *ref_37
        "412": *ref_38
        "500": *ref_39
        "501": *ref_40
      tags: *ref_41
      security: *ref_42
    patch:
      operationId: files-update
      summary: Изменить файлы
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FilesUpdateRequest"
      responses:
        "200":
          description: Успешный ответ
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FileGetResponse"
        "400": *ref_32
        "401": *ref_33
        "403": *ref_34
        "404": *ref_35
        "406": *ref_36
        "409": *ref_37
        "412": *ref_38
        "500": *ref_39
        "501": *ref_40
      tags: *ref_41
      security: *ref_42
    delete:
      operationId: files-delete
      summary: Удаление файла
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FilesDeleteRequest"
      responses:
        "200":
          description: Успешный ответ
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "400": *ref_32
        "401": *ref_33
        "403": *ref_34
        "404": *ref_35
        "406": *ref_36
        "409": *ref_37
        "412": *ref_38
        "500": *ref_39
        "501": *ref_40
      tags: *ref_41
      security: *ref_42
  /api/v2/file/copy:
    patch:
      operationId: files-copy
      summary: Скопировать файлы
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FilesCopyRequest"
      responses:
        "200":
          description: Успешный ответ
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FileGetResponse"
        "400": *ref_32
        "401": *ref_33
        "403": *ref_34
        "404": *ref_35
        "406": *ref_36
        "409": *ref_37
        "412": *ref_38
        "500": *ref_39
        "501": *ref_40
      tags: *ref_41
      security: *ref_42
  /api/v2/file/{fileId}:
    get:
      operationId: file-download
      summary: Скачивание файла
      parameters:
        - name: fileId
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: Успешный ответ
          content:
            video/mp4:
              encoding:
                video_mp4:
                  contentType: video/mp4
            image/jpeg:
              encoding:
                image_jpeg:
                  contentType: image/jpeg
            image/png:
              encoding:
                image_png:
                  contentType: image/png
        "400": *ref_32
        "401": *ref_33
        "403": *ref_34
        "404": *ref_35
        "406": *ref_36
        "409": *ref_37
        "412": *ref_38
        "500": *ref_39
        "501": *ref_40
      tags: *ref_41
      security: *ref_42
    post:
      operationId: file-get
      summary: Получить файл
      parameters:
        - name: fileId
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: Успешный ответ
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FileGetResponse"
        "400": *ref_32
        "401": *ref_33
        "403": *ref_34
        "404": *ref_35
        "406": *ref_36
        "409": *ref_37
        "412": *ref_38
        "500": *ref_39
        "501": *ref_40
      tags: *ref_41
      security: *ref_42
    patch:
      operationId: file-update
      summary: Изменить файл
      parameters:
        - name: fileId
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FileUpdateRequest"
      responses:
        "200":
          description: Успешный ответ
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FileGetResponse"
        "400": *ref_32
        "401": *ref_33
        "403": *ref_34
        "404": *ref_35
        "406": *ref_36
        "409": *ref_37
        "412": *ref_38
        "500": *ref_39
        "501": *ref_40
      tags: *ref_41
      security: *ref_42
    delete:
      operationId: file-delete
      summary: Удаление файла
      parameters:
        - name: fileId
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: Успешный ответ
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "400": *ref_32
        "401": *ref_33
        "403": *ref_34
        "404": *ref_35
        "406": *ref_36
        "409": *ref_37
        "412": *ref_38
        "500": *ref_39
        "501": *ref_40
      tags: *ref_41
      security: *ref_42
  /api/v2/file/{fileId}/preview:
    get:
      operationId: file-download-preview
      summary: Получить файл превью
      parameters:
        - name: fileId
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: Успешный ответ
          content:
            video/webm:
              encoding:
                video_webm:
                  contentType: video/webm
            image/jpeg:
              encoding:
                image_jpeg:
                  contentType: image/jpeg
        "400": *ref_32
        "401": *ref_33
        "403": *ref_34
        "404": *ref_35
        "406": *ref_36
        "409": *ref_37
        "412": *ref_38
        "500": *ref_39
        "501": *ref_40
      tags: *ref_41
      security: *ref_42
...
info:
  title: MyScreen backend API
  description: MyScreen backend API
  version: 2.41.29
  contact:
    name: Stanislav V Vyaliy
    url: https://wisekaa.dev
    email: stanislav@wisekaa.dev
...
servers: []
components:
  securitySchemes:
    bearer:
      scheme: bearer
      bearerFormat: JWT
      type: http
      description: Токен авторизации
      name: token
  schemas:
    BadRequestError:
      type: object
      properties:
        statusCode:
          type: number
          example: 400
        code:
          type: string
          example: server-error.10004
        message:
          type: string
          example: Bad request
      required:
        - statusCode
        - code
        - message
    UnauthorizedError:
      type: object
      properties:
        statusCode:
          type: number
          example: 401
        code:
          type: string
          example: server-error.10001
        message:
          type: string
          example: Unauthorized request
      required:
        - statusCode
        - code
        - message
    ForbiddenError:
      type: object
      properties:
        statusCode:
          type: number
          example: 403
        code:
          type: string
          example: server-error.10002
        message:
          type: string
          example: Forbidden
      required:
        - statusCode
        - code
        - message
    NotFoundError:
      type: object
      properties:
        statusCode:
          type: number
          example: 404
        code:
          type: string
          example: server-error.10005
        message:
          type: string
          example: Not Found
      required:
        - statusCode
        - code
        - message
    NotAcceptableError:
      type: object
      properties:
        statusCode:
          type: number
          example: 406
        code:
          type: string
          example: server-error.10000
        message:
          type: string
          example: Not Acceptable
      required:
        - statusCode
        - code
        - message
...
    ConflictError:
      type: object
      properties:
        statusCode:
          type: number
          example: 409
        code:
          type: string
          example: server-error.10000
        message:
          type: string
          example: Conflict exists
        data:
          $ref: "#/components/schemas/ConflictData"
      required:
        - statusCode
        - code
        - message
        - data
    PreconditionFailedError:
      type: object
      properties:
        statusCode:
          type: number
          example: 412
        code:
          type: string
          example: server-error.10002
        message:
          type: string
          example: User exists
      required:
        - statusCode
        - code
        - message
    InternalServerError:
      type: object
      properties:
        statusCode:
          type: number
          example: 500
        code:
          type: string
          example: server-error.10000
        message:
          type: string
          example: Server error
      required:
        - statusCode
        - code
        - message
    NotImplementedError:
      type: object
      properties:
        statusCode:
          type: number
          example: 501
        code:
          type: string
          example: server-error.10000
        message:
          type: string
          example: Server error
      required:
        - statusCode
        - code
        - message
    Status:
      type: string
      enum:
        - success
        - error
...
    SuccessResponse:
      type: object
      properties:
        status:
          description: Статус операции
          example: success
          $ref: "#/components/schemas/Status"
      required:
        - status
...
    LimitRequest:
      type: object
      properties:
        limit:
          type: number
          description: Лимит строк результатов
          example: 20
        page:
          type: number
          description: Страница результатов
          example: 1
        order:
          description: Порядок результатов
          example:
            createdAt: DESC
          allOf:
            - $ref: "#/components/schemas/LimitOrderRequest"
...
    FileResponse:
      type: object
      properties:
        id:
          type: string
          description: Идентификатор файла
          format: uuid
        name:
          type: string
          description: Имя файла
          example: foo.mp4
        extension:
          type: string
          description: Расширение файла
          example: mp4
        hash:
          type: string
          description: Hash файла
          example: 2b0439011a3a215ae1756bfc342e5bbc
        videoType:
          description: Тип файла
          example: video
          $ref: "#/components/schemas/VideoType"
        category:
          description: В какую категорию относить файл
          example: media
          $ref: "#/components/schemas/FileCategory"
        filesize:
          type: number
          description: Размер файла
          example: 210000
        duration:
          type: number
          description: Продолжительность видео
          example: 10
        width:
          type: number
          description: Размер по горизонтали
          example: 1024
        height:
          type: number
          description: Размер по вертикали
          example: 1024
        createdAt:
          format: date-time
          type: string
          description: Время создания
          example: "2021-01-01T10:00:00.147Z"
        updatedAt:
          format: date-time
          type: string
          description: Время изменения
          example: "2021-01-01T10:00:00.147Z"
        folder:
          description: Папка
          allOf:
            - $ref: "#/components/schemas/FolderResponse"
        monitors:
          description: Мониторы
          type: array
          items:
            $ref: "#/components/schemas/MonitorResponse"
        playlists:
          description: Плэйлисты
          type: array
          items:
            $ref: "#/components/schemas/PlaylistResponse"
      required:
        - id
        - name
        - extension
        - hash
        - videoType
        - category
        - filesize
        - duration
        - width
        - height
        - createdAt
        - updatedAt
        - folder
...
    FileUploadRequest:
      type: object
      properties:
        folderId:
          type: string
          description: Идентификатор папки
          format: uuid
        category:
          description: В какую категорию относить файл
          example: media
          $ref: "#/components/schemas/FileCategory"
        monitorId:
          type: string
          format: uuid
          description: Монитор
    FilePartialRequest:
      type: object
      properties:
        id:
          type: string
          description: Идентификатор файла
          format: uuid
        folderId:
          type: string
          description: Идентификатор папки
          format: uuid
        videoType:
          description: Тип файла
          example: video
          $ref: "#/components/schemas/VideoType"
        category:
          description: В какую категорию относить файл
          example: media
          $ref: "#/components/schemas/FileCategory"
        createdAt:
          description: Время создания
          example:
            - "2021-01-01"
            - "2021-12-31"
          type: array
          items:
            type: array
        updatedAt:
          description: Время изменения
          example:
            - "2021-01-01"
            - "2021-12-31"
          type: array
          items:
            type: array
      required:
        - folderId
    FilesGetRequest:
      type: object
      properties:
        where:
          description: Запрос
          allOf:
            - $ref: "#/components/schemas/FilePartialRequest"
        select:
          type: array
          description: Выбрать поля
          example: []
          items:
            type: string
            enum:
              - id
              - folderId
              - name
              - extension
              - hash
              - videoType
              - category
              - filesize
              - duration
              - width
              - height
              - createdAt
              - updatedAt
        scope:
          description: Рамки для запроса
          allOf:
            - $ref: "#/components/schemas/LimitRequest"
    FilesGetResponse:
      type: object
      properties:
        status:
          description: Статус операции
          example: success
          $ref: "#/components/schemas/Status"
        count:
          type: number
          description: Количество файлов
        data:
          description: Файлы
          type: array
          items:
            $ref: "#/components/schemas/FileResponse"
      required:
        - status
        - count
        - data
    FilesUploadResponse:
      type: object
      properties:
        status:
          description: Статус операции
          example: success
          $ref: "#/components/schemas/Status"
        count:
          type: number
          description: Количество файлов
        data:
          description: Файлы
          type: array
          items:
            $ref: "#/components/schemas/FileResponse"
      required:
        - status
        - count
        - data
...
    FilesUpdateRequest:
      type: object
      properties:
        files:
          description: Файлы
          type: array
          items:
            $ref: "#/components/schemas/FileUpdatesRequest"
      required:
        - files
    FileGetResponse:
      type: object
      properties:
        status:
          description: Статус операции
          example: success
          $ref: "#/components/schemas/Status"
        data:
          description: Файл
          allOf:
            - $ref: "#/components/schemas/FileResponse"
      required:
        - status
        - data
...
    FilesCopyRequest:
      type: object
      properties:
        toFolder:
          type: string
          description: Папка, куда копировать
          format: uuid
        files:
          description: Файлы
          type: array
          items:
            $ref: "#/components/schemas/FileCopyRequest"
      required:
        - toFolder
        - files
    FileUpdateRequest:
      type: object
      properties:
        folderId:
          type: string
          description: Идентификатор папки
          format: uuid
        name:
          type: string
          description: Имя файла
          example: foo.mp4
    FilesDeleteRequest:
      type: object
      properties:
        filesId:
          description: Файлы для удаления
          type: array
          items:
            type: string
            format: uuid
      required:
        - filesId
...